pyproject.toml
[tool.poetry]
name = "bloo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.1.0"
asyncpg = "^0.27.0"
aiolimiter = "1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
